requirements.txt
dapr-ext-fastapi
azure-storage-blob
fastapi~=0.115.14
dapr~=1.15.0
azure-core~=1.34.0
python-dotenv~=1.0.0
uvicorn~=0.23.2
torch~=2.3.0
opencensus~=0.11.2
opencensus-ext-azure~=1.1.0
opencensus-ext-logging~=0.1.1
opencensus-ext-fastapi~=0.1.0
opencensus-ext-requests~=0.1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
azure-monitor-opentelemetry-exporter
opentelemetry-semantic-conventions
opentelemetry-instrumentation-httpx
aiofiles
azure-core-tracing-opentelemetry
httpx